Result : refactor
